modify: refactor size control UI
